add Javadoc and fix grammar
